code rewrite - moved from separate connection to conn builder
