+++ lib/kafka/connection_builder.rb
@@ -1,3 +1,5 @@
+require 'kafka/sasl_gssapi_authenticator'
+
 module Kafka
   class ConnectionBuilder
     def initialize(client_id:, logger:, instrumenter:, connect_timeout:, socket_timeout:, ssl_context:, sasl_gssapi_principal:, sasl_gssapi_keytab:)
@@ -12,7 +14,7 @@ module Kafka
     end
 
     def build_connection(host, port)
-      Connection.new(
+      @connection = Connection.new(
         host: host,
         port: port,
         client_id: @client_id,
@@ -20,10 +22,23 @@ module Kafka
         socket_timeout: @socket_timeout,
         logger: @logger,
         instrumenter: @instrumenter,
-        ssl_context: @ssl_context,
-        sasl_gssapi_principal: @sasl_gssapi_principal,
-        sasl_gssapi_keytab: @sasl_gssapi_keytab
+        ssl_context: @ssl_context
       )
+      return @connection.socket unless sasl_gssapi_auth?
+      sasl_gssapi_authenticate
+    end
+
+    private
+
+    def sasl_gssapi_authenticate
+      auth = SaslGssapiAuthenticator.new(connection: @connection, logger: @logger,
+                                         sasl_gssapi_principal: @sasl_gssapi_principal,
+                                         sasl_gssapi_keytab: @sasl_gssapi_keytab)
+      @connection.socket if auth.authenticate!
+    end
+
+    def sasl_gssapi_auth?
+      !ssl_context && sasl_gssapi_principal && !sasl_gssapi_principal.empty?
     end
   end
 end
