+++ lib/kafka/protocol/join_group_request.rb
@@ -16,7 +16,7 @@ module Kafka
       end
 
       def api_key
-        ApiKey::JOIN_GROUP
+        11
       end
 
       def response_class
