+++ spec/protocol/sasl_handshake_request_spec.rb
@@ -1,10 +1,8 @@
-require 'kafka/protocol/api_key'
-
 describe Kafka::Protocol::SaslHandshakeRequest do
   describe "#api_key" do
     let(:request) { Kafka::Protocol::SaslHandshakeRequest.new('GSSAPI') }
     it 'expects correct api_key' do
-      expect(request.api_key).to eq Kafka::Protocol::ApiKey::SASL_HANDSHAKE
+      expect(request.api_key).to eq 17
     end
   end
